+++ test/api.ts
@@ -1,15 +1,24 @@
 import * as assert from "assert";
-import * as Postgrator from "../";
-
+import * as Postgrator from "../postgrator";
+import * as pg from "pg";
 import * as path from "path";
-const migrationDirectory = path.join(__dirname, "migrations");
-const pgUrl = "tcp://postgrator:postgrator@localhost:5432/postgrator";
 
-describe("TypeScript:API", function () {
+const migrationPattern = path.join(__dirname, "migrations/*");
+
+describe("TypeScript:API:execQuery", function () {
+  const client = new pg.Client({
+    host: "localhost",
+    port: 5432,
+    database: "postgrator",
+    user: "postgrator",
+    password: "postgrator",
+  });
+
   const postgrator = new Postgrator({
     driver: "pg",
-    migrationDirectory: migrationDirectory,
-    connectionString: pgUrl,
+    migrationPattern,
+    database: "postgrator",
+    execQuery: (query) => client.query(query),
   });
 
   const vStarted: Postgrator.Migration[] = [];
@@ -22,6 +31,10 @@ describe("TypeScript:API", function () {
   );
   postgrator.on("migration-started", (migration) => mStarted.push(migration));
   postgrator.on("migration-finished", (migration) => mFinished.push(migration));
+
+  before(async () => {
+    await client.connect();
+  });
 
   it("Migrates up to 003", async () => {
     const migrations: Postgrator.Migration[] = await postgrator.migrate("003");
@@ -50,7 +63,7 @@ describe("TypeScript:API", function () {
     const m = migrations[0];
     assert.strictEqual(m.version, 1);
     assert.strictEqual(m.action, "do");
-    assert.strictEqual(m.filename, "001.do.sql");
+    assert(m.filename.endsWith("001.do.sql"));
     assert(m.hasOwnProperty("name"));
   });
 
@@ -58,7 +71,6 @@ describe("TypeScript:API", function () {
     const patterngrator = new Postgrator({
       driver: "pg",
       migrationPattern: `${__dirname}/fail*/*`,
-      connectionString: pgUrl,
     });
     const migrationsByPattern: Postgrator.Migration[] =
       await patterngrator.getMigrations();
@@ -75,7 +87,8 @@ describe("TypeScript:API", function () {
     assert.strictEqual(migrations.length, 4, "4 migrations run");
   });
 
-  after((): Promise<Postgrator.QueryResult> => {
-    return postgrator.runQuery("DROP TABLE schemaversion");
+  after(async () => {
+    await postgrator.runQuery("DROP TABLE schemaversion");
+    await client.end();
   });
 });